feat:filter by device, renaming and docs
