+++ src/components/observationTable/ModalboxContent/MetadataCardsRowDialog.tsx
